Android Save and read bookmarks by URL not title #255

+++ src/org/kiwix/kiwixmobile/database/KiwixDatabase.java
@@ -20,6 +20,7 @@
 package org.kiwix.kiwixmobile.database;
 
 import android.content.Context;
+import android.util.Log;
 
 import com.yahoo.squidb.data.SquidCursor;
 import com.yahoo.squidb.data.SquidDatabase;
@@ -29,6 +30,8 @@ import com.yahoo.squidb.sql.Property;
 import com.yahoo.squidb.sql.Query;
 import com.yahoo.squidb.sql.Table;
 
+import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.ZimContentProvider;
 import org.kiwix.kiwixmobile.database.entity.BookDataSource;
 import org.kiwix.kiwixmobile.database.entity.BookDatabaseEntity;
 import org.kiwix.kiwixmobile.database.entity.Bookmarks;
@@ -36,6 +39,11 @@ import org.kiwix.kiwixmobile.database.entity.LibraryDatabaseEntity;
 import org.kiwix.kiwixmobile.database.entity.RecentSearch;
 import org.kiwix.kiwixmobile.database.entity.RecentSearchSpec;
 
+import java.io.BufferedReader;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -43,19 +51,23 @@ import java.util.List;
 
 public class KiwixDatabase extends SquidDatabase {
 
-  private static final int VERSION = 5;
+  private static final int VERSION = 6;
+  private Context context;
+
 
   public KiwixDatabase(Context context) {
     super(context);
+    this.context = context;
   }
 
-  @Override public String getName() {
+  @Override
+  public String getName() {
     return "Kiwix.db";
   }
 
   @Override
   protected Table[] getTables() {
-    return new Table[] {
+    return new Table[]{
         BookDatabaseEntity.TABLE,
         LibraryDatabaseEntity.TABLE,
         RecentSearch.TABLE,
@@ -63,17 +75,23 @@ public class KiwixDatabase extends SquidDatabase {
     };
   }
 
-  @Override protected boolean onUpgrade(SQLiteDatabaseWrapper db, int oldVersion, int newVersion) {
-    if (newVersion >= 3) {
+  @Override
+  protected boolean onUpgrade(SQLiteDatabaseWrapper db, int oldVersion, int newVersion) {
+    if (newVersion >= 3 && oldVersion < 3) {
       db.execSQL("DROP TABLE IF EXISTS recents");
       tryCreateTable(RecentSearch.TABLE);
     }
-    if (newVersion >= 4) {
+    if (newVersion >= 4 && oldVersion < 4) {
       tryCreateTable(Bookmarks.TABLE);
     }
-    if (newVersion >= 5) {
+    if (newVersion >= 5 && oldVersion < 5) {
       db.execSQL("DROP TABLE IF EXISTS book");
       tryCreateTable(BookDatabaseEntity.TABLE);
+    }
+    if (newVersion >= 6 && oldVersion < 6) {
+      db.execSQL("DROP TABLE IF EXISTS Bookmarks");
+      tryCreateTable(Bookmarks.TABLE);
+      BookmarksDao bookmarksDao = new BookmarksDao(this);
     }
     return true;
   }
